+++ ui/src/main/java/com/yaso/sample/InfoBase.java
@@ -1,0 +1,13 @@
+package com.yaso.sample;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.SerializationFeature;
+
+abstract class InfoBase {
+
+  public String toJson() throws Exception {
+    ObjectMapper mapper = new ObjectMapper();
+    mapper.enable(SerializationFeature.INDENT_OUTPUT);
+    return mapper.writeValueAsString(this);
+  }
+}